+++ androidApp/src/main/java/com/gdgnantes/devfest/androidapp/ui/screens/session/FeedbackForm.kt
@@ -1,27 +1,19 @@
 package com.gdgnantes.devfest.androidapp.ui.screens.session
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.gdgnantes.devfest.androidapp.BuildConfig
 import com.gdgnantes.devfest.androidapp.R
-import com.gdgnantes.devfest.androidapp.utils.getDateFromIso8601
 import com.gdgnantes.devfest.model.Session
-import com.gdgnantes.devfest.model.WebLinks
-import io.openfeedback.android.OpenFeedback
-import io.openfeedback.android.components.SessionFeedbackContainer
-import java.text.SimpleDateFormat
-import java.util.Locale
+import io.openfeedback.OpenFeedback
 
 @Composable
 fun FeedbackForm(
@@ -43,8 +35,7 @@ fun FeedbackForm(
             } else {
                 OpenfeedbackForm(
                     modifier = modifier,
-                    openFeedback = feedbackFormViewModel.openFeedback,
-                    openFeedbackFormId = openFeedbackFormId
+                    sessionId = openFeedbackFormId
                 )
             }
         }
@@ -52,46 +43,9 @@ fun FeedbackForm(
 }
 
 @Composable
-fun FallbackFeedbackForm(
-    modifier: Modifier = Modifier,
-    session: Session,
-    onFeedbackFormFallbackLinkClick: (String) -> Unit,
-) {
-    Box(
-        modifier =
-        modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-    ) {
-        OutlinedButton(
-            modifier = Modifier.align(Alignment.Center),
-            onClick = {
-                getFallbackUrlForSession(
-                    session
-                )?.run { onFeedbackFormFallbackLinkClick(this) }
-            }
-        ) {
-            Text(
-                text = stringResource(id = R.string.session_feedback_label)
-            )
-        }
-    }
-}
-
-private fun getFallbackUrlForSession(session: Session): String? {
-    val openFeedbackFormId = session.openFeedbackFormId ?: return null
-    return getDateFromIso8601(session.scheduleSlot.startDate)?.run {
-        val pattern = "yyyy-MM-dd"
-        val simpleDateFormat = SimpleDateFormat(pattern)
-        simpleDateFormat.format(this)
-    }?.run { "${WebLinks.OPENFEEDBACK_BASE_URL.url}/$this/$openFeedbackFormId" }
-}
-
-@Composable
 fun OpenfeedbackForm(
     modifier: Modifier = Modifier,
-    openFeedback: OpenFeedback,
-    openFeedbackFormId: String,
+    sessionId: String,
 ) {
     Column(
         modifier = modifier
@@ -102,13 +56,9 @@ fun OpenfeedbackForm(
             style = MaterialTheme.typography.headlineSmall
         )
 
-        SessionFeedbackContainer(
-            openFeedback = openFeedback,
-            sessionId = openFeedbackFormId,
-            language = Locale.getDefault().language,
-            modifier =
-            Modifier
-                .padding(horizontal = 8.dp)
+        OpenFeedback(
+            projectId = BuildConfig.OPEN_FEEDBACK_FIREBASE_PROJECT_ID,
+            sessionId = sessionId
         )
     }
 }
