+++ gradle/libs.versions.toml
@@ -26,7 +26,7 @@ kotlinxSerialization = "1.7.3"
 ksp = "2.0.20-1.0.25"
 material = "1.12.0"
 preference = "1.2.1"
-openfeedback = "0.0.7"
+openfeedback = "1.0.0-alpha.3"
 secretsPlugin = "2.0.1"
 spalshscreen = "1.0.1"
 timber = "5.0.1"
@@ -95,7 +95,8 @@ kotlinx-serialization-json = { group = "org.jetbrains.kotlinx", name = "kotlinx-
 
 material = { module = "com.google.android.material:material", version.ref = "material" }
 
-openfeedback = { module = "io.openfeedback:feedback-android-sdk-ui", version.ref = "openfeedback" }
+openfeedback-m3 = { module = "io.openfeedback:openfeedback-m3", version.ref = "openfeedback" }
+openfeedback-viewmodel = { module = "io.openfeedback:openfeedback-viewmodel", version.ref = "openfeedback" }
 
 timber = { module = "com.jakewharton.timber:timber", version.ref = "timber" }
